# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/cart/course/{courseID}": {
            "post": {
                "description": "AddController for Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "AddController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-cart_CartItem"
                        }
                    }
                }
            },
            "delete": {
                "description": "RemoveController for Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "RemoveController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-utils_DeleteResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "description": "ListController for Category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "ListController",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-array_category_Category"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateController for Category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "CreateController",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-category_Category"
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "delete": {
                "description": "DeleteController for Category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "DeleteController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-utils_DeleteResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateController for Category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "UpdateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-category_Category"
                        }
                    }
                }
            }
        },
...
        "/api/v1/certificates/claim/course/{courseID}": {
            "post": {
                "description": "ClaimController for Certificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificate"
                ],
                "summary": "ClaimController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-certificate_Certificate"
                        }
                    }
                }
            }
        },
        "/api/v1/certificates/course/{courseID}": {
            "get": {
                "description": "GetController for Certificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificate"
                ],
                "summary": "GetController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-certificate_Certificate"
                        }
                    }
                }
            }
        },
        "/api/v1/chapters/course/{courseID}": {
            "get": {
                "description": "ListController for Chapters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chapters"
                ],
                "summary": "ListController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-array_chapters_Chapter"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateController for Chapters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chapters"
                ],
                "summary": "CreateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courseID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chapters.CreateChapterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-chapters_Chapter"
                        }
                    }
                }
            }
        },
        "/api/v1/chapters/{id}": {
            "delete": {
                "description": "DeleteController for Chapters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chapters"
                ],
                "summary": "DeleteController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-utils_DeleteResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateController for Chapters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chapters"
                ],
                "summary": "UpdateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chapters.UpdateChapterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-chapters_Chapter"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cart.CartItem": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "course": {
                    "$ref": "#/definitions/models.CourseInfo"
                },
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "category.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "subcategories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/category.Category"
                    }
                }
            }
        },
        "category.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "parent_id": {
                    "type": "string"
                }
            }
        },
        "category.UpdateCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "certificate.Certificate": {
            "type": "object",
            "properties": {
                "course": {
                    "$ref": "#/definitions/models.CourseInfo"
                },
                "id": {
                    "type": "string"
                },
                "issued_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "chapters.Chapter": {
            "type": "object",
            "properties": {
                "chapter_no": {
                    "type": "integer"
                },
                "course_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_duration_seconds": {
                    "type": "integer"
                },
                "total_lectures": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "chapters.CreateChapterRequest": {
            "type": "object",
            "required": [
                "chapter_no",
                "title"
            ],
            "properties": {
                "chapter_no": {
                    "type": "integer",
                    "minimum": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                }
            }
        },
        "chapters.UpdateChapterRequest": {
            "type": "object",
            "properties": {
                "chapter_no": {
                    "type": "integer",
                    "minimum": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                }
            }
        },
...
        "models.CourseInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "utils.DeleteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "utils.SwaggerResponse-array_category_Category": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/category.Category"
                    }
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-array_chapters_Chapter": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chapters.Chapter"
                    }
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-cart_CartItem": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/cart.CartItem"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SwaggerResponse-category_Category": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/category.Category"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SwaggerResponse-certificate_Certificate": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/certificate.Certificate"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SwaggerResponse-chapters_Chapter": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/chapters.Chapter"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-utils_DeleteResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.DeleteResponse"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}